+++ NEWS.md
@@ -1,3 +1,11 @@
+jaatha 3.1.1 (in development)
+============
+
+* Fixes broken suppression of messages with `verbose` argument of the `jaatha` 
+  function.
+
+
+
 jaatha 3.1.0
 ============